# openapi.yaml
openapi: 3.0.3
info:
  title: Appliance Shop API - OpenAPI 3.0
  description: |-
    This is a Home Appliances Store Server based on the OpenAPI 3.0 specification. API for device management in Home Appliance Stores.

    Some useful links:
    - [The Appliance Shop repository](https://github.com/minhcopilot/Appliance-Shop-API)
    - [The source API definition for the Appliance Shop](https://github.com/minhcopilot/Appliance-Shop-API/openopenapi.yaml)

  contact:
    email: phanminh190502@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:9000
    description: Local server
...
paths:
...
  /products/suggestions:
    get:
      tags:
        - product
      summary: Get product suggestions by keyword
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /products:
    get:
      tags:
        - product
      summary: Get all products
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '204':
          description: No products
        '500':
          description: Internal Server Error

    post:
      tags:
        - product
      summary: Create a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                  format: float
                categoryId:
                  type: integer
                supplierId:
                  type: integer
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Product already exists
        '500':
          description: Internal Server Error
  /products/category/{categoryId}:
    get:
      tags:
        - product
      summary: Get products by category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '204':
          description: No products found for this category
        '500':
          description: Internal Server Error
  /products/search:
    get:
      tags:
        - product
      summary: Search products by keyword (name and description)
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '204':
          description: No products found
        '500':
          description: Internal Server Error
  /products/{id}:
    get:
      tags:
        - product
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not found
        '500':
          description: Internal Server Error

    patch:
      tags:
        - product
      summary: Update a product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                  format: float
                categoryId:
                  type: integer
                supplierId:
                  type: integer
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                removeFiles:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not found
        '500':
          description: Internal Server Error

    delete:
      tags:
        - product
      summary: Delete a product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
        '404':
          description: Not found
        '500':
          description: Internal Server Error
  /products/sorted-by-price:
    get:
      tags:
        - product
      summary: Get products sorted by price
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '204':
          description: No products
        '500':
          description: Internal Server Error
  /products/uploads/{id}:
    post:
      tags:
        - product
      summary: Upload images for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Success
        '500':
          description: Internal Server Error
  /products/upload/{id}:
    post:
      tags:
        - product
      summary: Upload an image for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Success
        '500':
          description: Internal Server Error

  #Customer
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Sample Product'
        price:
          type: number
          example: 100.0
        discount:
          type: number
          example: 10
        stock:
          type: integer
          example: 50
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT